requirements.txt
# Core Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.6
pydantic[email]>=2.4.0
pydantic-settings>=2.0.3
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
aiosmtplib>=3.0.0
Pillow>=10.0.0

# Development Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
hypothesis>=6.88.0
httpx>=0.25.0
